Store nested scopes, change main fn to "main" instead of ".main"
The idea is that scopes are always accessed top -> bottom and child scopes only need access to parent scopes, not vice versa.
Therefore it should be all right to simply only have access to the current scope (the scope with the most nested parents) in the interpreter and to let the Scope impl handle getting stuff from the other scopes.

Changed the name of the main function to "main" so that the user can define the main fn themselves, should they want to.